refactor(plcs): move generation config helper into generate_dataset/
PLCS の `prepare_generation_config` を `utils/config.py` から
`generate_dataset/config.py` へ移動し、BLCS (`generate_dataset/config.py`)
と配置を対称化する (issue #494, A案: 場所のみ対称化)。

- utils/__init__.py の再エクスポートを削除 (utils/ は pose_geometry 用に存続)
- chunk_manager.py / training/runner.py の import を直接参照へ更新
- 関数の責務・命名は変更なし
