+++ src/tasks/plcs/utils/__init__.py
@@ -1,5 +1,1 @@
 """Shared PLCS utilities."""
-
-from src.tasks.plcs.utils.config import prepare_generation_config
-
-__all__ = ["prepare_generation_config"]
